Cap rich to version 10.x
The latest version of rich includes breaking changes that affects ansible-lint.
A new version of ansible-lint has just been released with support for rich 11
but our Ansible infrastructure is on Molecule 2 which caps ansible-lint to 4.x.
Until we are ready to upgrade our infrastructure to Molecule 3, this should
allow our Ansible CI to keep testing roles and playbooks